chore(protocol-http): switch to HttpRequest static clone method
