+++ .changeset/poor-dolls-cheer.md
@@ -1,0 +1,8 @@
+---
+"@smithy/protocol-http": minor
+"@smithy/signature-v4": minor
+"@smithy/core": minor
+"@smithy/experimental-identity-and-auth": patch
+---
+
+switch to static HttpRequest clone method
