+++ packages/signature-v4/package.json
@@ -25,6 +25,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@smithy/is-array-buffer": "workspace:^",
+    "@smithy/protocol-http": "workspace:^",
     "@smithy/types": "workspace:^",
     "@smithy/util-hex-encoding": "workspace:^",
     "@smithy/util-middleware": "workspace:^",
@@ -34,7 +35,6 @@
   },
   "devDependencies": {
     "@aws-crypto/sha256-js": "5.2.0",
-    "@smithy/protocol-http": "workspace:^",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
