+++ packages/signature-v4/src/moveHeadersToQuery.ts
@@ -1,16 +1,14 @@
-import { HttpRequest, QueryParameterBag } from "@smithy/types";
-
-import { cloneRequest } from "./cloneRequest";
+import { HttpRequest } from "@smithy/protocol-http";
+import type { HttpRequest as IHttpRequest, QueryParameterBag } from "@smithy/types";
 
 /**
  * @private
  */
 export const moveHeadersToQuery = (
-  request: HttpRequest,
+  request: IHttpRequest,
   options: { unhoistableHeaders?: Set<string> } = {}
-): HttpRequest & { query: QueryParameterBag } => {
-  const { headers, query = {} as QueryParameterBag } =
-    typeof (request as any).clone === "function" ? (request as any).clone() : cloneRequest(request);
+): IHttpRequest & { query: QueryParameterBag } => {
+  const { headers, query = {} as QueryParameterBag } = HttpRequest.clone(request);
   for (const name of Object.keys(headers)) {
     const lname = name.toLowerCase();
     if (lname.slice(0, 6) === "x-amz-" && !options.unhoistableHeaders?.has(lname)) {
