+++ packages/signature-v4/src/prepareRequest.ts
@@ -1,14 +1,14 @@
-import { HttpRequest } from "@smithy/types";
+import { HttpRequest } from "@smithy/protocol-http";
+import type { HttpRequest as IHttpRequest } from "@smithy/types";
 
-import { cloneRequest } from "./cloneRequest";
 import { GENERATED_HEADERS } from "./constants";
 
 /**
  * @private
  */
-export const prepareRequest = (request: HttpRequest): HttpRequest => {
+export const prepareRequest = (request: IHttpRequest): IHttpRequest => {
   // Create a clone of the request object that does not clone the body
-  request = typeof (request as any).clone === "function" ? (request as any).clone() : cloneRequest(request);
+  request = HttpRequest.clone(request);
 
   for (const headerName of Object.keys(request.headers)) {
     if (GENERATED_HEADERS.indexOf(headerName.toLowerCase()) > -1) {
